services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
azure-cosmos>=4.7.0
msal>=1.29.0
azure-identity>=1.17.0
azure-keyvault-secrets>=4.8.0
openai>=1.40.0
stripe>=10.5.0
requests>=2.32.0
opencensus-ext-azure>=1.1.13
# Container Apps Jobs SDK -- used by tools/trigger_aca_job.py
# Install: pip install azure-mgmt-appcontainers
azure-mgmt-appcontainers>=3.0.0
pytest>=8.3.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.5.0
mypy>=1.11.0
